requirements.txt
cbor2==5.4.6
click==8.1.3
crytic-compile==0.3.0
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
prettytable==3.6.0
pycryptodome==3.17
slither-analyzer==0.9.2
wcwidth==0.2.6
Werkzeug==2.2.3
